Following uncomplete list of changes:

- Added Greedy meshing
- Begin of Biome Refining
- Introduced max threading (greatly improve performance)
- Refactored and restructured a lot of code
- Structure Generation is now part of the TerrainGenerator and works per-chunk.

+++ Voxel Engine/Assets/Scripts/Voxel Engine/WorldGen/Noise/DomainWarping.cs
@@ -9,20 +9,20 @@ namespace Voxel_Engine.WorldGen.Noise
         public int amplitudeX = 20, amplitudeY = 20;
         
         
-        public float GenerateDomainNoise(int x, int z, NoiseSettings defaultNoiseSettings)
+        public float GenerateDomainNoise(float x, float z, NoiseSettings defaultNoiseSettings)
         {
             Vector2 domainOffset = GenerateDomainOffset(x, z);
             return MyNoise.OctavePerlin(x + domainOffset.x, z + domainOffset.y, defaultNoiseSettings);
         }
 
-        public Vector2 GenerateDomainOffset(int x, int z)
+        public Vector2 GenerateDomainOffset(float x, float z)
         {
             var noiseX = MyNoise.OctavePerlin(x, z, noiseDomainX) * amplitudeX;
             var noiseY = MyNoise.OctavePerlin(x, z, noiseDomainY) * amplitudeY;
             return new Vector2(noiseX, noiseY);
         }
 
-        public Vector2Int GenerateDomainOffsetInt(int x, int z)
+        public Vector2Int GenerateDomainOffsetInt(float x, float z)
         {
             return Vector2Int.RoundToInt(GenerateDomainOffset(x, z));
         }
